web: add support for systemd socket activation
The web.Server type now supports systemd socket activation. This allows a service to be started on demand when a connection is made to its socket, which is managed by systemd.

To enable socket activation, set the Addr field of web.Server to the special value "sd-socket:<name>", where <name> corresponds to the FileDescriptorName set in the systemd socket unit file.

To support this, the systemd package has been extended with a new Socket function that retrieves named listeners passed by systemd. The package has also been reorganized to separate the sd_notify logic from the new socket activation logic.

This feature is only available on Linux.